+++ app/src/main/java/com/moroom/android/ui/nav/map/MapFragment.kt
@@ -1,4 +1,4 @@
-package com.moroom.android.ui.navui.map
+package com.moroom.android.ui.nav.map
 
 import android.app.Activity
 import android.content.Intent
