+++ app/src/main/java/com/moroom/android/ui/write/WriteActivity.java
@@ -28,7 +28,7 @@ import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
 import com.moroom.android.ui.adapter.write.CheckedTextViewAdapter;
-import com.moroom.android.ui.navui.MainActivity;
+import com.moroom.android.ui.nav.MainActivity;
 import com.moroom.android.ui.search.SearchActivity;
 
 import java.util.ArrayList;
