+++ app/src/main/res/layout/fragment_home.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="viewModel"
-            type="com.moroom.android.ui.navui.home.HomeViewModel" />
+            type="com.moroom.android.ui.nav.home.HomeViewModel" />
     </data>
 
     <RelativeLayout
